# project-dam-backend/dist/dist/swagger.json
{
    "openapi": "3.0.0",
    "components": {
...
        "schemas": {
            "mongoose.Types.ObjectId": {
                "type": "string"
            },
            "IUser": {
                "properties": {
                    "_id": {
                        "$ref": "#/components/schemas/mongoose.Types.ObjectId"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "guest",
                            "technician",
                            "admin"
                        ]
                    },
                    "isActive": {
                        "type": "boolean"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "required": [
                    "_id",
                    "name",
                    "email",
                    "password",
                    "role",
                    "isActive",
                    "createdAt",
                    "updatedAt"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "CreateUserRequest": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "guest",
                            "technician",
                            "admin"
                        ]
                    }
                },
                "required": [
                    "name",
                    "email",
                    "password",
                    "role"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "UpdateUserRequest": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "guest",
                            "technician",
                            "admin"
                        ]
                    },
                    "isActive": {
                        "type": "boolean"
                    }
                },
                "type": "object",
                "additionalProperties": false
            },
...
        },
        "securitySchemes": {
            "jwt": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "info": {
        "title": "project-dam-backend",
        "version": "1.0.0",
        "license": {
            "name": "ISC"
        },
        "contact": {}
    },
    "paths": {
        "/users": {
            "get": {
                "operationId": "GetUsers",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "pages": {
                                            "type": "number",
                                            "format": "double"
                                        },
                                        "total": {
                                            "type": "number",
                                            "format": "double"
                                        },
                                        "users": {
                                            "items": {
                                                "$ref": "#/components/schemas/IUser"
                                            },
                                            "type": "array"
                                        }
                                    },
                                    "required": [
                                        "pages",
                                        "total",
                                        "users"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "description": "Get all users (Admin only)",
                "tags": [
                    "Users"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "page",
                        "required": false,
                        "schema": {
                            "default": 1,
                            "format": "double",
                            "type": "number"
                        }
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "required": false,
                        "schema": {
                            "default": 10,
                            "format": "double",
                            "type": "number"
                        }
                    }
                ]
            },
            "post": {
                "operationId": "CreateUser",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IUser"
                                }
                            }
                        }
                    }
                },
                "description": "Create a new user (Admin only)",
                "tags": [
                    "Users"
                ],
                "security": [
                    {
                        "jwt": [
                            "admin"
                        ]
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserRequest"
                            }
                        }
                    }
                }
            }
        },
        "/users/profile": {
            "get": {
                "operationId": "GetProfile",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IUser"
                                }
                            }
                        }
                    }
                },
                "description": "Get current user profile",
                "tags": [
                    "Users"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": []
            }
        },
        "/users/{userId}": {
            "get": {
                "operationId": "GetUser",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IUser"
                                }
                            }
                        }
                    }
                },
                "description": "Get user by ID",
                "tags": [
                    "Users"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "userId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            },
            "put": {
                "operationId": "UpdateUser",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IUser"
                                }
                            }
                        }
                    }
                },
                "description": "Update user (Admin only)",
                "tags": [
                    "Users"
                ],
                "security": [
                    {
                        "jwt": [
                            "admin"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "userId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserRequest"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "DeleteUser",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "description": "Delete user (Admin only)",
                "tags": [
                    "Users"
                ],
                "security": [
                    {
                        "jwt": [
                            "admin"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "userId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
...
    },
    "servers": [
        {
            "url": "/"
        }
    ]
}